typo

+++ src/manager/source_driver_ros2.hpp
@@ -282,7 +282,7 @@ inline sensor_msgs::msg::PointCloud2 SourceDriver::ToRosMsg(const LidarDecodedFr
   sensor_msgs::PointCloud2Iterator<float> iter_intensity_(ros_msg, "intensity");
   sensor_msgs::PointCloud2Iterator<uint16_t> iter_ring_(ros_msg, "ring");
   sensor_msgs::PointCloud2Iterator<double> iter_timestamp_(ros_msg, "timestamp");
-  sensor_msgs::PointCloud2Iterator<double> iter_range_(ros_msg, "range");
+  sensor_msgs::PointCloud2Iterator<float> iter_range_(ros_msg, "range");
   for (size_t i = 0; i < points_number; i++)
   {
     LidarPointXYZIRT point = pPoints[i];
